+++ src/primitive/str/ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: rlutt <rlutt@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/12/01 18:41:09 by rlutt             #+#    #+#             */
-/*   Updated: 2017/03/06 18:24:54 by rlutt            ###   ########.fr       */
+/*   Updated: 2018/04/13 13:41:36 by dauie            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,8 +18,8 @@ size_t	ft_strlcat(char *dst, const char *src, int len)
 	int		dlen;
 	int		i;
 
-	slen = ft_strlen(src);
-	dlen = ft_strlen(dst);
+	slen = (int)ft_strlen(src);
+	dlen = (int)ft_strlen(dst);
 	i = 0;
 	if (len - 1 <= dlen)
 		return ((size_t)(slen + len));
